refactor(connections): share tunnel rewrite and reconnect logic between SSH and Cloudflare (#1285)
