+++ TablePro/Core/Database/DatabaseManager+SSH.swift
@@ -6,8 +6,6 @@
 //
 
 import Foundation
-import os
-import TableProPluginKit
 
 // MARK: - SSH Tunnel Helper
 
@@ -73,38 +71,7 @@ extension DatabaseManager {
             totpPeriod: sshConfig.totpPeriod
         )
 
-        // Adapt SSL config for tunnel: SSH already authenticates the server,
-        // remote environment and aren't readable locally, so strip them and
-        // use at least .preferred so libpq negotiates SSL when the server
-        // requires it (SSH already authenticates the server itself).
-        var tunnelSSL = connection.sslConfig
-        if tunnelSSL.isEnabled {
-            if tunnelSSL.verifiesCertificate {
-                tunnelSSL.mode = .required
-            }
-            tunnelSSL.caCertificatePath = ""
-            tunnelSSL.clientCertificatePath = ""
-            tunnelSSL.clientKeyPath = ""
-        }
-
-        var effectiveFields = connection.additionalFields
-        if connection.usePgpass {
-            effectiveFields["pgpassOriginalHost"] = connection.host
-            effectiveFields["pgpassOriginalPort"] = String(connection.port)
-        }
-
-        return DatabaseConnection(
-            id: connection.id,
-            name: connection.name,
-            host: "127.0.0.1",
-            port: tunnelPort,
-            database: connection.database,
-            username: connection.username,
-            type: connection.type,
-            sshConfig: SSHConfiguration(),
-            sslConfig: tunnelSSL,
-            additionalFields: effectiveFields
-        )
+        return tunneledConnection(from: connection, localPort: tunnelPort)
     }
 
     // MARK: - SSH Tunnel Recovery
@@ -114,46 +81,10 @@ extension DatabaseManager {
     /// when both the keepalive death callback and the wake-from-sleep handler fire
     /// for the same connection.
     func handleSSHTunnelDied(connectionId: UUID) async {
-        guard let session = activeSessions[connectionId],
-              !recoveringConnectionIds.contains(connectionId) else { return }
-
-        recoveringConnectionIds.insert(connectionId)
-        defer { recoveringConnectionIds.remove(connectionId) }
-
-        Self.logger.warning("SSH tunnel died for connection: \(session.connection.name)")
-
-        // Stop health monitor before retrying to prevent stale pings during reconnect
-        await stopHealthMonitor(for: connectionId)
-
-        // Disconnect the stale driver and invalidate it so connectToSession
-        // creates a fresh connection instead of short-circuiting on driver != nil
-        activeSessions[connectionId]?.driver?.disconnect()
-        updateSession(connectionId) { session in
-            session.driver = nil
-            session.status = .connecting
-        }
-
-        let maxRetries = 10
-        for retryCount in 0..<maxRetries {
-            let delay = ExponentialBackoff.delay(for: retryCount + 1, maxDelay: 120)
-            Self.logger.info("SSH reconnect attempt \(retryCount + 1)/\(maxRetries) in \(delay)s for: \(session.connection.name)")
-            try? await Task.sleep(nanoseconds: UInt64(delay * 1_000_000_000))
-
-            do {
-                try await connectToSession(session.connection)
-                Self.logger.info("Successfully reconnected SSH tunnel for: \(session.connection.name)")
-                return
-            } catch {
-                Self.logger.warning("SSH reconnect attempt \(retryCount + 1) failed: \(error.localizedDescription)")
-            }
-        }
-
-        Self.logger.error("All SSH reconnect attempts failed for: \(session.connection.name)")
-
-        // Mark as error and release stale cached data
-        updateSession(connectionId) { session in
-            session.status = .error("SSH tunnel disconnected. Click to reconnect.")
-            session.clearCachedData()
-        }
+        await recoverDeadTunnel(
+            connectionId: connectionId,
+            kind: "SSH",
+            disconnectedMessage: String(localized: "SSH tunnel disconnected. Click to reconnect.")
+        )
     }
 }
